Refactored key-parsing from EZConfig
