+++ XMonad/Keys/Events.hs
@@ -1,0 +1,1 @@
+module XMonad.Keys.Events where
