+++ XMonad/Keys/Weird.hs
@@ -1,0 +1,10 @@
+{-
+The soon to be renamed "Weird" module.
+
+This code contains everything I need to make my own weirdo QMK-like keymappings.
+
+-}
+
+module XMonad.Keys.Weird where
+
+import XMonad
